+++ R/disk.R
@@ -121,6 +121,80 @@ ps__disk_usage_format_posix <- function(paths, l) {
   d
 }
 
+#' System-wide disk I/O counters
+#'
+#' Returns a data.frame of system-wide disk I/O counters.
+#'
+#' Includes the following non-NA fields for all supported platforms:
+#' * `read_count`: number of reads
+#' * `write_count`: number of writes
+#' * `read_bytes`: number of bytes read
+#' * `write_bytes`: number of bytes written
+#'
+#' And for only some platforms:
+#' * `read_time`: time spent reading from disk (in milliseconds)
+#' * `write_time`: time spent writing to disk (in milliseconds)
+#' * `busy_time`: time spent doing actual I/Os (in milliseconds)
+#' * `read_merged_count`: number of merged reads (see iostats doc)
+#' * `write_merged_count`: number of merged writes (see iostats doc)
+#'
+#' @return A data frame of one row per disk of I/O stats, with columns
+#' `name`, `read_count` `read_merged_count` `read_bytes`, `read_time`,
+#' `write_count`, `write_merged_count`, `write_bytes` `write_time`, and
+#' `busy_time`.
+#'
+#' @family disk functions
+#' @export
+#' @examplesIf ps:::ps_os_name() %in% c("LINUX", "WINDOWS") && !ps:::is_cran_check()
+#' ps_disk_io_counters()
+ps_disk_io_counters <- function() {
+  os <- ps_os_name()
+  tab <- if (os == "LINUX") {
+    ps__disk_io_counters_linux()
+  } else if (os == "WINDOWS") {
+    ps__disk_io_counters_windows()
+  } else {
+    ps__disk_io_counters_macos()
+  }
+
+  class(tab) <- c("tbl", "data.frame")
+  tab
+}
+
+ps__disk_io_counters_windows <- function() {
+  l <- .Call(ps__disk_io_counters)
+  disk_info <- data_frame(
+    name = l[[1]],
+    read_count = l[[2]],
+    read_merged_count = NA,
+    read_bytes = l[[3]],
+    read_time = l[[4]],
+    write_count = l[[5]],
+    write_merged_count = NA,
+    write_bytes = l[[6]],
+    write_time = l[[7]],
+    busy_time = NA
+  )
+
+  disk_info[disk_info$name != "",]
+}
+
+ps__disk_io_counters_macos <- function() {
+  tab <- not_null(.Call(ps__disk_io_counters))
+  data_frame(
+    name = names(tab),
+    read_count = map_dbl(tab, "[[", 1),
+    read_merged_count = NA,
+    read_bytes = map_dbl(tab, "[[", 3),
+    read_time = map_dbl(tab, "[[", 5),
+    write_count = map_dbl(tab, "[[", 2),
+    write_merged_count = NA,
+    write_bytes = map_dbl(tab, "[[", 4),
+    write_time = map_dbl(tab, "[[", 6),
+    busy_time = NA
+  )
+}
+
 #' File system information for files
 #'
 #' @param paths A path or a vector of paths. `ps_fs_info()` returns