+++ tests/testthat/test-disk.R
@@ -6,3 +6,27 @@ test_that("ps_fs_info", {
     ps_fs_info(c("/", "~", "."))
   )
 })
+
+test_that("disk_io", {
+  result <- ps_disk_io_counters()
+
+  # Check structure
+  expect_named(
+    result,
+    c(
+      "read_bytes",
+      "write_bytes",
+      "read_count",
+      "write_count",
+      "read_merged_count",
+      "read_time",
+      "write_merged_count",
+      "write_time",
+      "busy_time",
+      "name"
+    ),
+    ignore.order = TRUE
+  )
+  expect_type(result, "list")
+  expect_s3_class(result, "data.frame")
+})